simplify send.
Splits `lib/send` into two pieces:

- `wrap-send` wraps an IPC implementations. It reformats two arguments (eventName, data) into a single data object, and sends via the supplied send callback.

- `send-to-parent` wraps `process.send` using the above wrapper function. It also provides the "must be used via CLI" message.

`wrap-send` is only used by `fork.js` and the newly introduced `send-to-parent`.

`send-to-parent` can be used anywhere in the child process to send data to the parent (currently `index.js`, and `test-worker.js`).
