+++ lib/fork.js
@@ -6,7 +6,7 @@ var objectAssign = require('object-assign');
 var Promise = require('bluebird');
 var debug = require('debug')('ava');
 var AvaError = require('./ava-error');
-var doSend = require('./send');
+var wrapSend = require('./wrap-send');
 
 if (fs.realpathSync(__filename) !== __filename) {
 	console.warn(
@@ -50,14 +50,14 @@ module.exports = function (file, opts) {
 	var relFile = path.relative('.', file);
 
 	var exiting = false;
-	var send = function (ps, name, data) {
+	var send = wrapSend(function (data) {
 		if (!exiting) {
 			// This seems to trigger a Node bug which kills the AVA master process, at
 			// least while running AVA's tests. See
 			// <https://github.com/novemberborn/_ava-tap-crash> for more details.
-			doSend(ps, name, data);
+			ps.send(data);
 		}
-	};
+	});
 
 	var testResults = [];
 	var results;
@@ -86,7 +86,7 @@ module.exports = function (file, opts) {
 		ps.on('results', function (data) {
 			results = data;
 			data.tests = testResults;
-			send(ps, 'teardown');
+			send('teardown');
 		});
 
 		ps.on('exit', function (code, signal) {
@@ -106,7 +106,7 @@ module.exports = function (file, opts) {
 		});
 
 		ps.on('no-tests', function (data) {
-			send(ps, 'teardown');
+			send('teardown');
 
 			var message = 'No tests found in ' + relFile;
 
@@ -120,13 +120,13 @@ module.exports = function (file, opts) {
 
 	// teardown finished, now exit
 	ps.on('teardown', function () {
-		send(ps, 'exit');
+		send('exit');
 		exiting = true;
 	});
 
 	// uncaught exception in fork, need to exit
 	ps.on('uncaughtException', function () {
-		send(ps, 'teardown');
+		send('teardown');
 	});
 
 	ps.stdout.on('data', function (data) {
@@ -144,13 +144,13 @@ module.exports = function (file, opts) {
 	};
 
 	promise.send = function (name, data) {
-		send(ps, name, data);
+		send(name, data);
 
 		return promise;
 	};
 
 	promise.exit = function () {
-		send(ps, 'init-exit');
+		send('init-exit');
 
 		return promise;
 	};
@@ -164,12 +164,12 @@ module.exports = function (file, opts) {
 
 	promise.run = function (options) {
 		if (isReady) {
-			send(ps, 'run', options);
+			send('run', options);
 			return promise;
 		}
 
 		ps.on('stats', function () {
-			send(ps, 'run', options);
+			send('run', options);
 		});
 
 		return promise;
