+++ lib/test-worker.js
@@ -62,7 +62,7 @@ sourceMapSupport.install({
 
 var currentlyUnhandled = require('currently-unhandled')(); // eslint-disable-line import/order
 var serializeError = require('./serialize-error');
-var send = require('./send');
+var send = require('./send-to-parent');
 var throwsHelper = require('./throws-helper');
 var installPrecompiler = require('require-precompiled'); // eslint-disable-line import/order
 var cacheDir = opts.cacheDir;
